+++ sdk/src/main/java/com/mapbox/maps/renderer/egl/EGLCore.kt
@@ -1,14 +1,23 @@
 package com.mapbox.maps.renderer.egl
 
+import android.os.Handler
+import android.os.Looper
 import android.view.Surface
+import com.mapbox.maps.MapView
 import com.mapbox.maps.logE
 import com.mapbox.maps.logI
+import com.mapbox.maps.logW
+import com.mapbox.maps.renderer.RendererError
+import com.mapbox.maps.renderer.RendererSetupErrorListener
+import java.util.*
+import java.util.concurrent.locks.ReentrantLock
 import javax.microedition.khronos.egl.*
+import kotlin.concurrent.withLock
 
 /**
  * Core EGL state (display, context, config).
  *
- * The EGLContext must only be attached to one thread at a time.  This class is not thread-safe.
+ * The EGLContext must only be attached to one thread at a time. This class is not thread-safe.
  *
  * Inspired by [Grafika](https://github.com/google/grafika/blob/master/app/src/main/java/com/android/grafika/gles/EglCore.java)
  */
@@ -24,45 +33,57 @@ internal class EGLCore(
 
   internal val eglNoSurface: EGLSurface = EGL10.EGL_NO_SURFACE
 
+  /**
+   * For user convenience we will deliver [RendererSetupErrorListener] on main thread.
+   */
+  private val mainHandler = Handler(Looper.getMainLooper())
+  /**
+   * We require error accumulation as renderer is created when [MapView] is created and
+   * some errors could already be triggered on render thread
+   * before user will call [MapView.addRendererSetupErrorListener] -
+   * in that case accumulated errors will still be delivered to the user.
+   */
+  private val accumulatedRendererErrorList = LinkedList<RendererError>()
+  private val rendererSetupErrorListenerSet = LinkedList<RendererSetupErrorListener>()
+  private val rendererErrorLock = ReentrantLock()
+
   fun prepareEgl(): Boolean {
     egl = EGLContext.getEGL() as EGL10
     eglDisplay = egl.eglGetDisplay(EGL10.EGL_DEFAULT_DISPLAY)
     if (eglDisplay == EGL10.EGL_NO_DISPLAY) {
-      logE(TAG, "Unable to get EGL14 display")
-      return false
+      // no need to report error upstream here due to https://www.khronos.org/registry/EGL/sdk/docs/man/html/eglGetDisplay.xhtml
+      logW(TAG, "Unable to get default display, eglInitialize will most likely fail shortly.")
     }
-
     val versions = IntArray(2)
     if (!egl.eglInitialize(eglDisplay, versions)) {
-      logE(TAG, "eglInitialize failed")
+      checkEglErrorAndNotify("eglInitialize")
       return false
     }
-
     EGLConfigChooser(translucentSurface, antialiasingSampleCount).chooseConfig(egl, eglDisplay)?.let {
       eglConfig = it
     } ?: run {
+      notifyListeners(RendererError.NoValidEglConfigFound)
       return false
     }
-
     val context = egl.eglCreateContext(
       eglDisplay,
       eglConfig,
       sharedContext,
       intArrayOf(EGL_CONTEXT_CLIENT_VERSION, 2, EGL10.EGL_NONE)
     )
-    val contextCreated = checkEglErrorNoException("eglCreateContext")
-    if (!contextCreated) {
-      return false
-    }
+    checkEglErrorAndNotify("eglCreateContext")
     eglContext = context
     // Confirm with query.
     val values = IntArray(1)
-    egl.eglQueryContext(
+    val eglQueryContextSuccess = egl.eglQueryContext(
       eglDisplay,
       eglContext,
       EGL_CONTEXT_CLIENT_VERSION,
       values
     )
+    if (!eglQueryContextSuccess) {
+      checkEglErrorAndNotify("eglQueryContext")
+    }
     logI(TAG, "EGLContext created, client version ${values[0]}")
     return true
   }
@@ -86,7 +107,7 @@ internal class EGLCore(
   }
 
   /**
-   * Destroys the specified surface.  Note the EGLSurface won't actually be destroyed if it's
+   * Destroys the specified surface. Note the EGLSurface won't actually be destroyed if it's
    * still current in a context.
    */
   fun releaseSurface(eglSurface: EGLSurface) {
@@ -107,9 +128,8 @@ internal class EGLCore(
       surface,
       surfaceAttribs
     )
-    val eglWindowCreated = checkEglErrorNoException("eglCreateWindowSurface")
-    if (!eglWindowCreated || eglSurface == null) {
-      logE(TAG, "Surface was null")
+    val eglWindowCreatedError = checkEglErrorAndNotify("eglCreateWindowSurface")
+    if (eglWindowCreatedError != null || eglSurface == null) {
       return eglNoSurface
     }
     return eglSurface
@@ -126,9 +146,8 @@ internal class EGLCore(
       eglConfig,
       surfaceAttribs
     )
-    checkEglErrorNoException("eglCreatePbufferSurface")
-    if (eglSurface == null) {
-      logE(TAG, "Offscreen surface was null")
+    val eglCreatePbufferSurfaceError = checkEglErrorAndNotify("eglCreatePbufferSurface")
+    if (eglCreatePbufferSurfaceError != null || eglSurface == null) {
       return eglNoSurface
     }
     return eglSurface
@@ -139,10 +158,8 @@ internal class EGLCore(
    */
   fun makeNothingCurrent(): Boolean {
     if (!egl.eglMakeCurrent(eglDisplay, EGL10.EGL_NO_SURFACE, EGL10.EGL_NO_SURFACE, EGL10.EGL_NO_CONTEXT)) {
-      val eglMakeNothingCurrentSuccess = checkEglErrorNoException("eglMakeNothingCurrent")
-      if (!eglMakeNothingCurrentSuccess) {
-        return false
-      }
+      checkEglErrorAndNotify("makeNothingCurrent")
+      return false
     }
     return true
   }
@@ -159,7 +176,7 @@ internal class EGLCore(
       logI(TAG, "NOTE: makeCurrent w/o display")
     }
     if (!egl.eglMakeCurrent(eglDisplay, eglSurface, eglSurface, eglContext)) {
-      logE(TAG, "eglMakeCurrent failed")
+      checkEglErrorAndNotify("eglMakeCurrent")
       return false
     }
     return true
@@ -176,16 +193,68 @@ internal class EGLCore(
     return EGL10.EGL_SUCCESS
   }
 
-  /**
-   * Checks for EGL errors.
-   */
-  private fun checkEglErrorNoException(msg: String): Boolean {
+  internal fun addRendererStateListener(listener: RendererSetupErrorListener) {
+    rendererErrorLock.withLock {
+      rendererSetupErrorListenerSet.add(listener)
+      if (accumulatedRendererErrorList.isNotEmpty()) {
+        val errorListCopy = LinkedList(accumulatedRendererErrorList)
+        accumulatedRendererErrorList.clear()
+        mainHandler.post {
+          errorListCopy.forEach { error ->
+            listener.onError(error)
+          }
+        }
+      }
+    }
+  }
+
+  internal fun removeRendererStateListener(listener: RendererSetupErrorListener) {
+    rendererErrorLock.withLock {
+      rendererSetupErrorListenerSet.remove(listener)
+    }
+  }
+
+  internal fun clearRendererStateListeners() {
+    rendererErrorLock.withLock {
+      accumulatedRendererErrorList.clear()
+      rendererSetupErrorListenerSet.clear()
+    }
+  }
+
+  private fun checkEglErrorAndNotify(functionName: String): Int? {
+    val eglError = checkEglError(functionName)
+    if (eglError != null) {
+      val mappedError = if (eglError == EGL10.EGL_BAD_ALLOC)
+        RendererError.OutOfMemory
+      else
+        RendererError.EglError(eglError)
+      logE(TAG, "EGL error ${mappedError.eglErrorCode} occurred for $functionName!")
+      notifyListeners(mappedError)
+    }
+    return eglError
+  }
+
+  private fun checkEglError(msg: String): Int? {
     val error = egl.eglGetError()
     if (error != EGL10.EGL_SUCCESS) {
       logE(TAG, msg + ": EGL error: 0x${Integer.toHexString(error)}")
-      return false
+      return error
     }
-    return true
+    return null
+  }
+
+  private fun notifyListeners(error: RendererError) {
+    rendererErrorLock.withLock {
+      if (rendererSetupErrorListenerSet.isEmpty()) {
+        accumulatedRendererErrorList.add(error)
+      } else {
+        mainHandler.post {
+          rendererSetupErrorListenerSet.forEach {
+            it.onError(error)
+          }
+        }
+      }
+    }
   }
 
   companion object {